BUILD: Reject GNU Make 3.80 and older
GNU Make 3.80 is the default version on macOS Tiger, and it shouldn't be
used because it has various parsing problems and bugs.

The .FEATURES variable was added in GNU Make 3.81, and that's the
easiest and most reliable way of detecting this version.

We know that GNU Make 3.81 is OK since that's still what Apple ships
by default on modern macOS (because of GPLv3).

(cherry picked from commit 4c90e20b8d81d260c9e54176ce929c3c96a041db)